+++ src/main/java/core/SynthLogic/Effects/DestructionEffect.java
@@ -1,13 +1,12 @@
 package core.SynthLogic.Effects;
 
 import core.Constants.ConstantValues;
-import org.jtransforms.fft.FloatFFT_1D;
 
 import java.util.Random;
 
 public class DestructionEffect implements EffectRack {
     @Override
-    public float[] applyEffect(float[] mixBuffer) {
+    public double[] applyEffect(double[] mixBuffer) {
         for (int i = 0; i < mixBuffer.length; i++) {
             mixBuffer[i] += Math.pow(mixBuffer[i], randomDistorter(20));
             mixBuffer[i] = softClip(mixBuffer[i]);
@@ -19,22 +18,21 @@ public class DestructionEffect implements EffectRack {
         return mixBuffer;
     }
 
-    private float softClip(float sample) {
-        if (sample > 1.0f) {
-            return 1.0f - (1.0f / (1.0f + sample));
-        } else if (sample < -1.0f) {
-            return -1.0f + (1.0f / (1.0f - sample));
+    private double softClip(double sample) {
+        if (sample > 1.0) {
+            return 1.0 - (1.0 / (1.0 + sample));
+        } else if (sample < -1.0) {
+            return -1.0 + (1.0 / (1.0 - sample));
         } else {
             return sample;
         }
     }
 
-    private float randomDistorter(float amount) {
+    private double randomDistorter(double amount) {
         Random randomGenerator = new Random();
 
-        float distortionValue = randomGenerator.nextFloat();
-        float randomDistortion = distortionValue / amount;
+        double distortionValue = randomGenerator.nextDouble();
+        double randomDistortion = distortionValue / amount;
         return randomDistortion;
     }
-
 }
